Generate configuration cache problems when build logic crosses project boundaries when isolated projects is enabled and running a tooling API action.
